+++ egui/src/demos/widgets.rs
@@ -133,19 +133,14 @@ impl Widgets {
 
         ui.horizontal(|ui| {
             ui.label("Single line text input:");
-            let response = ui.add(
-                TextEdit::new(&mut self.single_line_text_input)
-                    .multiline(false)
-                    .id_source("single line"),
-            );
-
+            let response = ui.text_edit_singleline(&mut self.single_line_text_input);
             if response.lost_kb_focus {
                 // The user pressed enter.
             }
         });
 
         ui.label("Multiline text input:");
-        ui.add(TextEdit::new(&mut self.multiline_text_input).id_source("multiline"));
+        ui.text_edit_multiline(&mut self.multiline_text_input);
 
         ui.separator();
         super::toggle_switch::demo(ui, &mut self.toggle_switch);